pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stt-realtime-demo"
version = "0.1.0"
description = "Real-time STT Demo with Deepgram"
authors = [
    {name = "park.byunghyun", email = "byunghyun@illuni.com"},
]
requires-python = "==3.12.*"
dependencies = [
    "starlette>=0.37.2",
    "uvicorn[standard]>=0.29.0",
    "streamlit>=1.33.0",
    "websockets>=12.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "deepgram-sdk>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "jupyter>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["server", "client"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
